Fixes more compiler warnings.
Most of these warnings are again constructor initializer-list
re-ordering warnings from MinGW. There are also a few type comparison
warnings fixed, and some un-used variables removed. Further inspection
actually also fixes a memory leak when loading invalid RGB[A] textures.

+++ src/CoreGraphics/Shader.cpp
@@ -82,7 +82,7 @@ bool zShader::LoadFromRaw(const string_t& string)
     GLint length = string.length();
     GL(glShaderSource(shader, 1, &src, &length));
 
-    ZEN_ASSERT(length == string.length());
+    ZEN_ASSERT(static_cast<size_t>(length) == string.length());
 
     GL(glCompileShader(shader));
     GL(glGetShaderiv(shader, GL_COMPILE_STATUS, &error_code));
